feat(url): added multiple url parsers
